requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
pdfplumber
python-multipart
passlib[bcrypt]
rsa
playwright
loguru
httpx
pytest
python-docx
mem0ai
qdrant-client
sentence-transformers
openai
